% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle_info_data.R
\docType{data}
\name{cycle_info}
\alias{cycle_info}
\title{Information on public bike rental stations}
\format{An object of class \code{data.frame} with 1460 rows and 15 columns.}
\source{
\href{https://data.seoul.go.kr/}{Seoul Open Data Plaza}
}
\usage{
data(cycle_info)
}
\description{
Data that provides information on the location and surroundings of public bicycle rental stations
}
\examples{
data(cycle_info)
}
\keyword{datasets}
