% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle_log_c_14_fac_data.R
\docType{data}
\name{cycle_log_c_14_fac}
\alias{cycle_log_c_14_fac}
\title{Data for public bicycle model training}
\format{An object of class \code{data.frame} with 8856360 rows and 16 columns.}
\source{
\href{https://data.seoul.go.kr/}{Seoul Open Data Plaza}
}
\usage{
data(cycle_log_c_14_fac)
}
\description{
Training Data for Predicting Demand Model for Public Bicycles
}
\examples{
data(cycle_log_c_14_fac)
}
\keyword{datasets}
