% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_area_data.R
\docType{data}
\name{interest_area}
\alias{interest_area}
\title{derivative variable}
\format{An object of class \code{data.frame} with 222 rows and 1 columns.}
\source{
\href{https://data.seoul.go.kr/}{Seoul Open Data Plaza}
}
\usage{
data(interest_area)
}
\description{
Data for the generation of derived variables
}
\examples{
data(interest_area)
}
\keyword{datasets}
