% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_time_bicycle.R
\name{real_time_bicycle}
\alias{real_time_bicycle}
\title{Real-time Public Bicycle Rental station Status}
\usage{
real_time_bicycle(id = 0, key)
}
\arguments{
\item{id}{bicycle rental station id}

\item{key}{Seoul Open Data Plaza api key}
}
\value{
number of bicycles remaining in real time
}
\description{
This function shows the number of bicycles remaining in the public bike rental station in real time.
}
\examples{
  real_time_bicycle(358, key)
}
