% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_time_crawling.R
\name{real_time_crawling}
\alias{real_time_crawling}
\title{real-time information crawling}
\usage{
real_time_crawling(id, weather_time, key)
}
\arguments{
\item{id}{bicycle rental station id}

\item{weather_time}{Specify how many hours ago weather data should be collected}

\item{key}{Seoul Open Data Plaza api key}
}
\value{
explanatory variables for the public bicycle demand prediction model in real time
}
\description{
A function that collects the required explanatory variables for the public bicycle demand prediction model in real time.
}
\examples{
  real_time_crawling(358, 1, key)
}
