% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_time_plot.R
\name{real_time_plot}
\alias{real_time_plot}
\title{Public Bicycle Rentalable Density Function}
\usage{
real_time_plot(id, key)
}
\arguments{
\item{id}{bicycle rental station id}

\item{key}{Seoul Open Data Plaza api key}
}
\value{
Rentalable Density Function
}
\description{
Function showing the density function that can be rented when visiting the rental office in one hour
}
\examples{
  real_time_plot(358, key)
}
