% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_time_prediction.R
\name{real_time_prediction}
\alias{real_time_prediction}
\title{Predicting the Demand of Public Bicycles for an hour}
\usage{
real_time_prediction(id, key)
}
\arguments{
\item{id}{bicycle rental station id}

\item{key}{Seoul Open Data Plaza api key}
}
\value{
Predictive demand that will occur over an hour
}
\description{
A function to predict the demand that will occur over an hour at a particular rental office.
}
\examples{
  real_time_prediction(358, key)
}
