% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_time_prob.R
\name{real_time_prob}
\alias{real_time_prob}
\title{Public Bicycle Rentability}
\usage{
real_time_prob(id, key)
}
\arguments{
\item{id}{bicycle rental station id}

\item{key}{Seoul Open Data Plaza api key}
}
\value{
Public Bicycle Rentability
}
\description{
A function that calculates the probability of public bicycle rental when visiting a specific rental office after one hour.
}
\examples{
  real_time_prob(358, key)
}
